requirements.txt
pypdf==6.1.3
torch==2.9.0
accelerate==1.11.0
bitsandbytes==0.48.2
ctransformers==0.2.27
sentence_transformers==5.1.2
faiss_cpu==1.12.0
chainlit==2.9.0
huggingface_hub==0.36.0
python-dotenv==1.2.1
llama-cpp-python==0.3.16
langchain==1.0.5
langchain-community==0.4.1
langchain-core==1.0.4
langchain-classic==1.0.0
gradio==5.49.1